review fixes

+++ src/main/java/club/dnd5/portal/repository/MenuRepository.java
@@ -3,5 +3,5 @@ package club.dnd5.portal.repository;
 import club.dnd5.portal.model.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MenuRepository extends JpaRepository<Menu, Integer> {
+public interface MenuRepository extends JpaRepository<Menu, Short> {
 }
